Add support for CloudFormation events and response
CloudFormation works differently to the other events in that when an
event is received by a Lambda function, the response is sent to a
pre-signed URL given in the Event.

A new 'cfn' module exists as a result of this. Tests are
included for making sure things like Content-Type are never accidentally
sent as these would be responded to with 403 Forbidden.

A wrapper is included to allow customers to make Custom Resources
without having to create a lot of boilerplate.

Certain properties must be copied to a response and helper function is
included for this.

Sample README added for a simple Echo resource.