fix slots
If you set Slots as "map[string]string", when you send them delegating the action, amazon lex suppose that the next slots have been filled and goes directly to the "FulfillmentCodeHook". So, the type of the slots should be "map[string]interface{}" to send back the slots values as (nil), not empty string ("").