Resolve TestSESEventMarshaling failure
Using omitempty like this will prevent the inclusion of fields with
empty strings.

There are possibly scenarios where adding omitempty to the existing
FunctionARN and InvocationType fields would change the behaviour of
systems that expect the fields to be present and assigned an empty
string. I've decided that the consistency of the fields when
Unmarshaling is a bigger priority here though, as the chances of real
world applications depending on this behaviour seems pretty remote to
me.