Extend events.SimpleEmailReceiptAction fields (#211)
* Extend events.SimpleEmailReceiptAction fields

The action object contains different data depending on the type.
Previously only the fields relevant to the Lambda type were included,
preventing use of the SimpleEmailReceiptAction for other types.

* Update ses.go

* ReceiptAction FunctionArn -> FunctionARN
* ReceiptAction json:"smtpReplayCode" -> json:"smtpReplyCode"

* Resolve TestSESEventMarshaling failure

Using omitempty like this will prevent the inclusion of fields with
empty strings.

There are possibly scenarios where adding omitempty to the existing
FunctionARN and InvocationType fields would change the behaviour of
systems that expect the fields to be present and assigned an empty
string. I've decided that the consistency of the fields when
Unmarshaling is a bigger priority here though, as the chances of real
world applications depending on this behaviour seems pretty remote to
me.

* Extend TestSESEventMarshaling

This test only covered an SES lambda event, and did not cover other
types of SES events. Adding SES-S3 and SES-SNS events here is improving
the test coverage.

* Document events.SimpleEmailReceiptAction