S3 size is optional
The `size` field can be empty. It just so happens marshaling uses Go's default which is `0`, so this doesn't show up as an error. But if you inspect the payload it can be empty.

Though not directly of concern to the go lib, without this it is causing https://github.com/LegNeato/aws-lambda-events/issues/11.